Enable auto-save-mode via a variable instead of a function
If you use the variable, it should check to make sure that the buffer
visits a file before enabling auto-save. If you run the function, it
enables it in the current mode also, which I guess is the scratch
buffer. This is a thing I kind of just lived with for a long time
without really caring to solve.